requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
xgboost==2.1.3
imbalanced-learn==0.13.0
joblib==1.4.2
kafka-python==2.0.2
pyspark==3.5.4
pyarrow==18.1.0
boto3==1.35.90
pytest==8.3.4
